refactor: changes folder structure for resumable gallery example
Adds comments to some other components to provide a better understanding
of what they do.
